requirements.txt
Flask==3.0.0
pymssql>=2.2.11
bcrypt==4.1.2
PyYAML==6.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow>=10.2.0
regex==2023.10.3
gunicorn==21.2.0
PyJWT==2.8.0
pyOpenSSL==23.3.0